feat: Added a "prepare" button and a density property
This is used to send to an external cocktail preparing API that will
only receive:

- tap_number
- quantity
- ingredient density

It is configured via a single endpoint authenticated by api key, and has
a new "login" page.

Once the app opens for the first time, it'll request a mixology OTP
(wich you'll get by scanning a mixologist QR code), configure the
credentials and the endpoint, and show a welcome page with instructions
for the device.