feat: Renamed project
Kept the original links and credit to `drinkable` main app.

Keeping this as a fork as its main goal is to provide the app for the
device, so the UX will be oriented aournd that, wich is something not
useful for the main app.

Changed names, made contact form optional, made API url configurable (in
case someone wants to setup its own mixology api).